requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
billiard==3.6.3.0
blessings==1.7
bpython==0.19
certifi==2020.6.20
chardet==3.0.4
curtsies==0.3.4
decorator==4.4.2
Django==3.1.1
django-debug-toolbar==3.1.1
django-extensions==3.0.9
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
jsonpickle==1.4.1
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
middleware==1.2.3
mysqlclient==2.0.1
parso==0.7.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psutil==5.7.2
ptpython==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pydot==1.4.1
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
